# Conflicts:
#	src/NuGetGallery.Core/Infrastructure/TableErrorLog.cs
#	src/NuGetGallery/Web.config

+++ src/NuGetGallery.Core/Auditing/AuditedEntities/AuditedPackageRegistration.cs
@@ -8,7 +8,7 @@ namespace NuGetGallery.Auditing.AuditedEntities
     public class AuditedPackageRegistration
     {
         public string Id { get; private set; }
-        public int DownloadCount { get; private set; }
+        public long DownloadCount { get; private set; }
         public int Key { get; private set; }
         public bool IsVerified { get; private set; }
