+++ src/NuGetGallery/Services/DownloadCountObjectMaterializedInterceptor.cs
@@ -2,9 +2,7 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
-using System.Collections.Generic;
 using NuGet.Services.Entities;
-using NuGet.Services.Logging;
 
 namespace NuGetGallery
 {
@@ -37,7 +35,7 @@ namespace NuGetGallery
                 ? NuGetVersionFormatter.Normalize(package.Version)
                 : package.NormalizedVersion;
 
-            int downloadCount;
+            long downloadCount;
             if (_downloadCountService.TryGetDownloadCountForPackage(package.PackageRegistration.Id, packageNormalizedVersion, out downloadCount))
             {
                 if (downloadCount < package.DownloadCount)
@@ -56,7 +54,7 @@ namespace NuGetGallery
                 return;
             }
 
-            int downloadCount;
+            long downloadCount;
             if (_downloadCountService.TryGetDownloadCountForPackageRegistration(packageRegistration.Id, out downloadCount))
             {
                 if (downloadCount < packageRegistration.DownloadCount)
